Added blog to other dashboards

+++ resources/views/blogs/index.blade.php
@@ -29,10 +29,13 @@
 
     <center>
         <div class="d-flex gap-2 justify-content-center">
-            <button class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#editPostModal" onclick="editPost({{$blog}})">Edit</button>
-            <button class="btn btn-danger" onclick="deletePost('{{route('artist.blog.delete',$blog->id)}}')">Delete</button>
+            @if(auth()->user()->hasRole('admin') || (auth()->user()->hasRole('artist') && auth()->user()->id == $blog->artist_id))
+                <button class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#editPostModal" onclick="editPost({{ $blog }})">Edit</button>
+                <button class="btn btn-danger" onclick="deletePost('{{ route('artist.blog.delete', $blog->id) }}')">Delete</button>
+            @endif
         </div>
     </center>
+
 
     <div class="mt-4">
         <p class="text-justify text-secondary text-black">
@@ -42,7 +45,7 @@
 
     <div class="mt-3 d-flex justify-content-between align-items-center">
         <div>
-            <form action="{{ route('artist.blog.like', $blog->id) }}" method="POST">
+            <form action="{{ route(auth()->user()->getRoleNames()->first().'.blog.like', $blog->id) }}" method="POST">
                 @csrf
                 <button class="btn btn-outline-primary d-flex align-items-center gap-2">
                     <svg xmlns="http://www.w3.org/2000/svg" width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="lucide lucide-thumbs-up"><path d="M7 10v12"/><path d="M15 5.88 14 10h5.83a2 2 0 0 1 1.92 2.56l-2.33 8A2 2 0 0 1 17.5 22H4a2 2 0 0 1-2-2v-8a2 2 0 0 1 2-2h2.76a2 2 0 0 0 1.79-1.11L12 2a3.13 3.13 0 0 1 3 3.88Z"/></svg>
@@ -82,7 +85,7 @@
 
         <div class="mt-4">
             <h2 class="h5 font-weight-bold text-dark mb-3">Leave a Comment</h2>
-            <form action="{{route('artist.blog.comment',$blog->id)}}" method="POST">
+            <form action="{{route(auth()->user()->getRoleNames()->first().'.blog.comment',$blog->id)}}" method="POST">
                 @csrf
                 <div class="form-group">
                     <textarea
